Enable glint on ssr-web
