--- packages/ssr-web/app/components/profile-logo/index.hbs
@@ -1,16 +1,0 @@
-{{#if (first-char @name)}}
-  <div
-    class={{cn 'profile-logo' profile-logo--lg=(eq @size 'large')}}
-    style={{
-      css-var
-      profile-logo-background=(or @logoBackground 'var(--boxel-blue)')
-      profile-logo-text-color=(or @logoTextColor 'var(--boxel-light)')
-    }}
-    data-test-profile-logo
-    data-test-profile-logo-background={{@logoBackground}}
-    data-test-profile-logo-text-color={{@logoTextColor}}
-    ...attributes
-  >
-    {{capitalize (first-char @name)}}
-  </div>
-{{/if}}
