+++ packages/ssr-web/package.json
@@ -39,6 +39,9 @@
     "@embroider/webpack": "^1.8.3",
     "@glimmer/component": "^1.1.2",
     "@glimmer/tracking": "^1.1.2",
+    "@glint/core": "^0.9.4",
+    "@glint/environment-ember-loose": "^0.9.4",
+    "@glint/environment-ember-template-imports": "^0.9.4",
     "@hyperbola/svgo-loader": "0.1.2",
     "@percy/cli": "^1.0.0-beta.70",
     "@percy/ember": "^3.0.0",
@@ -116,6 +119,7 @@
     "ember-sinon-qunit": "^6.0.0",
     "ember-source": "~4.6.0",
     "ember-svg-jar": "^2.3.3",
+    "ember-template-imports": "^3.1.1",
     "ember-template-lint": "^3.16.0",
     "ember-window-mock": "^0.8.1",
     "eslint": "^7.32.0",
@@ -172,7 +176,8 @@
     "web3": "^1.7.1"
   },
   "resolutions": {
-    "ember-test-selectors": "^6.0.0"
+    "ember-test-selectors": "^6.0.0",
+    "@babel/generator": "~7.16.0"
   },
   "config": {
     "access": "public"
